+++ CHANGELOG.md
@@ -36,3 +36,15 @@ Before: `alert(message)` (old one can be called using `mralert(message)`
 - New feature! `QuickLaunch`, a Spotlight Search (like) app launcher for MrServer. (+ customizable keyboard shortcut. default: meta+k)
 - Added New API! (`await generateApps()`) which gives an array of the system and third-party apps. example: `[{"title": "app name", "icon": "icon", "subtitle": "System App", "category": "system", "function": "appfunction()"}]`
 - Made folder and file icons in File Manager adaptive to theme
+- Added New API! `moveWindowTo(x, y, animate = false)`
+- New feature! Registering app settings and users modify app settings from the MrServer Settings App.
+- Added New API! `registerAppSettings(id, config)`:
+    `id` is the App ID.
+    `config` is an `Object` with `title`, `controls`.
+        `controls` is an object array with these for each object:
+            `id`, `type` (range, number, select, checkbox, text (default)), `label`, `min` (number,range), `max` (number,range), `step` (number,range), `default`, `placeholder` (text), `readonly` (text), `value` (select)
+
+- Added New API! `getAppSetting(appId, settingId, fallback = null)`
+- Added New API! `setAppSetting(appId, settingId, value)`
+- Added new feature! Widgets are here!
+- Added New API! `createWidget(id, html, width, height, x, y, updateFn, options = {})`